refactor(tcp-plug-socket)!: merge sync and async version in one package
this allows to reuse some common parts
and maybe further reducing boilerplate
