--- smarthome/crates/libs/tcp-plug-socket-async/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-edition           = "2021"
-license.workspace = true
-name              = "tcp-plug-socket-async"
-version           = "0.1.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tokio = { version = "1.36.0", features = [
-    "io-util",
-    "macros",
-    "net",
-    "rt-multi-thread",
-    "time",
-] }
-
-anyhow = "1.0.79"
-# -- App Libs
-smarthome = { path = "../../libs/smarthome" }
-derive_more = "0.99.17"
-[lints]
-workspace = true
