+++ smarthome/crates/libs/tcp-plug-socket/Cargo.toml
@@ -7,9 +7,16 @@ version           = "0.1.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-anyhow    = "1.0.79"
+anyhow = "1.0.79"
+derive_more = "0.99.17"
 thiserror = { workspace = true }
-
+tokio = { version = "1.36.0", features = [
+    "io-util",
+    "macros",
+    "net",
+    "rt-multi-thread",
+    "time",
+] }
 # -- App Libs
 smarthome = { path = "../../libs/smarthome" }
