+++ smarthome/crates/libs/tcp-plug-socket/examples/scs.rs
