+++ smarthome/crates/libs/tcp-plug-socket/examples/tpi.rs
