+++ smarthome/crates/libs/tcp-plug-socket/src/impl_sync.rs
@@ -1,17 +1,12 @@
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::net::{TcpStream, ToSocketAddrs};
+use std::{
+    io::{Read, Write},
+    net::{TcpStream, ToSocketAddrs},
+};
 
-pub use self::command::Command;
-pub use self::responses::{OddResponse, Response};
-use error::WrongResponse;
-use smarthome::devices::{Gauge, SendCommandAsync};
+use crate::{error::Error, Command, OddResponse, Response};
+use smarthome::devices::{Gauge, SendCommand};
 
 type Result<T> = core::result::Result<T, Error>;
-type Error = Box<dyn std::error::Error>; // For early dev.
-
-mod command;
-mod error;
-mod responses;
 
 /// Odd TCP Smart Plug socket client- connects to smart plug via `TcpStream`
 /// let's pretend for some reason `HW` implementation can't handle floats
@@ -33,23 +28,23 @@ impl Gauge<Power> for TcpPlugOddSocket {
     }
 }
 
-impl SendCommandAsync<Command> for TcpPlugOddSocket {
+impl SendCommand<Command> for TcpPlugOddSocket {
     type R = Result<Response>;
     #[must_use]
-    async fn send_command(&mut self, command: Command) -> Self::R {
-        let odd_resp = self.send_command(u8::from(command)).await?;
+    fn send_command(&mut self, command: Command) -> Self::R {
+        let odd_resp = self.send_command(u8::from(command))?;
         self.update_state(odd_resp);
         Ok(self.convert_response(odd_resp))
     }
 }
 
-impl SendCommandAsync<u8> for TcpPlugOddSocket {
+impl SendCommand<u8> for TcpPlugOddSocket {
     type R = Result<OddResponse>;
-    async fn send_command(&mut self, raw_command: u8) -> Self::R {
-        self.stream.write_all(&[raw_command]).await?;
+    fn send_command(&mut self, raw_command: u8) -> Self::R {
+        self.stream.write_all(&[raw_command])?;
         // reading OddResponse
         let mut buffer = [0u8; 4];
-        self.stream.read_exact(&mut buffer).await?;
+        self.stream.read_exact(&mut buffer)?;
         // now we need convert it to Response type
         Ok((&buffer).into())
     }
@@ -61,25 +56,25 @@ impl TcpPlugOddSocket {
     /// # Errors
     ///
     /// This function will return an error if connection or calibration fails
-    pub async fn new<T: ToSocketAddrs + Send>(plug_addr: T) -> Result<Self> {
-        let stream = TcpStream::connect(plug_addr).await?;
+    pub fn new(plug_addr: impl ToSocketAddrs) -> Result<Self> {
+        let stream = TcpStream::connect(plug_addr)?;
         let calibrated = Self {
             stream,
             delimiter: 1.0,
             cached_pu: 0.0,
         }
-        .calibrate()
-        .await?;
+        .calibrate()?;
         Ok(calibrated)
     }
-    async fn calibrate(mut self) -> Result<Self> {
-        let raw: u8 = Command::Reserved { command_id: 42 }.into();
-        let res = self.send_command(raw).await?;
+    fn calibrate(mut self) -> Result<Self> {
+        const CALIBRATE_CMD: Command = Command::Reserved { command_id: 42 };
+        let raw: u8 = CALIBRATE_CMD.into();
+        let res = self.send_command(raw)?;
         match res {
             OddResponse::Reserved(buf) if buf[0] == 42u8 => {
                 self.delimiter = u16::from_be_bytes(buf[1..].try_into().unwrap_or([0, 1])).into();
             }
-            _ => return Err(WrongResponse::new(format!("{res}")).into()),
+            _ => return Err(Error::WrongResponse(CALIBRATE_CMD, res)),
         };
         Ok(self)
     }
@@ -87,14 +82,8 @@ impl TcpPlugOddSocket {
         f32::from(power) / self.delimiter
     }
     fn update_state(&mut self, r: OddResponse) {
-        match r {
-            OddResponse::Power(power) => self.cached_pu = self.scale_power(power),
-            OddResponse::Ok
-            | OddResponse::Enabled
-            | OddResponse::Disabled
-            | OddResponse::Retry
-            | OddResponse::MaxPower(_)
-            | OddResponse::Reserved(_) => {}
+        if let OddResponse::Power(power) = r {
+            self.cached_pu = self.scale_power(power);
         }
     }
