+++ smarthome/justfile
@@ -19,6 +19,6 @@ coverage:
 watch:
   cargo watch --clear --postpone -x 'clippy --lib' -x 'test --lib -- --show-output' -x 'run'
 tcp_plug:
-  cargo run --package tcp-plug-socket --example tcp_plugsocket_imitator
+  cargo run --package tcp-plug-socket --example tpi
 tcp_command_sender:
-  cargo run --package tcp-plug-socket --example simple_command_sender
+  cargo run --package tcp-plug-socket --example scs
